As an experiment, made all coordinates of a COORD_TYPE (macro constant). When it is "long", flash size goes from 90 to 95%, and RAM size from 72 to 74%. Not sure if timings got bad - needs to be tested.

+++ stacker.ino
@@ -31,7 +31,7 @@
    h1.1 [s0.10,s0.12,s0.14, s0.08a]: Second row keypad pin moved from 10 to 7. Pin 10 left free (for hardware SPI). Display's pin SCE (CE / chip select) disconnected from pin 7.
                 Instead, display SCE pin is soldered to the ground via 10k (pulldown) resistor.
    h1.2 [s1.00]: LCD reset pin (RST) disconnected from Arduino; instead it is now hardware controlled via RC delay circuit (R=47k, C=0.1uF, connected to VCC=+3.3V).
-                 Arduino pin 6 is now used to control the second relay (+ diod + R=33 Ohm), for camera autofocus.
+                  Arduino pin 6 is now used to control the second relay (+ diod + R=33 Ohm), for camera autofocus.
 */
 #include <EEPROM.h>
 #include <math.h>
@@ -86,7 +86,7 @@ void setup() {
   for (int i = 0; i < N_PARAMS; i++)
   {
     float fsteps = MM_PER_FRAME[i] / MM_PER_MICROSTEP;
-    short steps = nintMy(fsteps);
+    short steps = (short)nintMy(fsteps);
     if (steps < 20)
       MM_PER_FRAME[i] = ((float)steps) * MM_PER_MICROSTEP;
   }
